Refactor constants directory
