+++ frontend/layouts/default.vue
@@ -67,7 +67,7 @@
 </template>
 
 <script>
-import { LS_USER_DATA } from '../constants';
+import { LS_USER_DATA } from '../constants/constants';
 export default {
   data() {
     return {
