+++ frontend/pages/signup.vue
@@ -66,7 +66,7 @@
 </template>
 
 <script>
-import { LS_USER_DATA } from '../constants';
+import { LS_USER_DATA } from '../constants/constants';
 
 const EMAIL_VALIDATION_REGEX = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
 const PASSWORD_MIN_LENGTH = 6;
